feat(HttpReact): - Adds the HttpReact object, which can be used to make requests imperatively. It has a method for each HTTP verb, like 'get', 'post', etc. It also has a 'extend' method, similar to 'Axios.create'
