+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "http-react",
-  "version": "2.6.6",
+  "version": "2.6.7",
   "description": "React hooks for data fetching",
   "main": "dist/index.js",
   "scripts": {
