+++ src/internal/index.ts
@@ -1,3 +1,4 @@
+'use client'
 import { createContext, useContext } from 'react'
 
 import { CacheStoreType, FetchContextType } from '../types'
